Moved the shopping cart functions higher up on the project file tree
