+++ client/shopping-cart-functions.js
